fix: cement v2 gRPC protocol with correctness and design fixes
Critical fixes:
- Add Exited event (with exit_code) to ContainerStreamResponse so clients
  can distinguish clean exit from connection loss
- Restore RestartPolicy to StartContainerRequest (regression from v1)
- Make ConnectToWiFiRequest.password optional (required field breaks open networks)
- WiFi mutations (Forget/SetPriority/Reorder) now identify networks by UUID
  instead of SSID to avoid ambiguity when multiple profiles share an SSID;
  add UUID-based methods to NetworkManager and nmcli implementation
- Restore storage_medium to GetDeviceInfoResponse (was dropped from v1)
- Restore mcp_port to AppContainer in shared.proto (was dropped from v1)
- Remove size/sha256 from FileSyncChmod (chmod changes permissions, not content)

Design fixes:
- Replace success/error_message pattern with gRPC status codes throughout
  WiFi and Audio services
- Change ScanBluetoothPeripherals from bidi-streaming to server-streaming
  (empty client messages served no purpose)
- Replace free-form phase string in UpdateOSResponse.Progress with UpdatePhase enum
- Replace optional int32 min_severity in StreamLogsRequest with LogSeverity enum
  matching OTel SeverityNumber ranges
- Make ListContainerStats streaming (one stat per response) to match ListContainers
- Remove redundant sample_rate/channels from AudioChunk (fixed for stream lifetime,
  set once in StreamAudioRequest)
- Provision IDs promoted from int32 to int64

Regenerate Go stubs and update all v2 service implementations.

+++ go/internal/agent/services/provisioning_service_v2.go
@@ -35,8 +35,8 @@ func (s *ProvisioningServiceV2) IsProvisioned(ctx context.Context, _ *agentpbv2.
 		ResponseType: &agentpbv2.IsProvisionedResponse_Provisioned{
 			Provisioned: &agentpbv2.ProvisionedResponse{
 				CloudHost:      p.CloudHost,
-				OrganizationId: p.OrganizationId,
-				AssetId:        p.AssetId,
+				OrganizationId: int64(p.OrganizationId),
+				AssetId:        int64(p.AssetId),
 			},
 		},
 	}, nil
@@ -44,10 +44,10 @@ func (s *ProvisioningServiceV2) IsProvisioned(ctx context.Context, _ *agentpbv2.
 
 func (s *ProvisioningServiceV2) StartProvisioning(ctx context.Context, req *agentpbv2.StartProvisioningRequest) (*agentpbv2.StartProvisioningResponse, error) {
 	if _, err := s.v1.StartProvisioning(ctx, &agentpb.StartProvisioningRequest{
-		OrganizationId:  req.OrganizationId,
+		OrganizationId:  int32(req.OrganizationId),
 		EnrollmentToken: req.EnrollmentToken,
 		CloudHost:       req.CloudHost,
-		AssetId:         req.AssetId,
+		AssetId:         int32(req.AssetId),
 	}); err != nil {
 		return nil, err
 	}
